+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.15] - 2022-05-09
+
+### Added
+
+- Unit tests on cars and users domain
+
+### Changed
+
+### Removed
+
+### Fixed
+
 ## [0.0.14] - 2022-05-09
 
 ### Added